[FIX] #69 : product serving response, number Int -> Double로 변경

+++ client/data/src/main/java/com/healthc/data/model/remote/product/ProductResponse.kt
@@ -32,7 +32,7 @@ data class ProductResponse(
 }
 
 data class ProductServingResponse(
-    val number: Int,
+    val number: Double,
     val size: Double?,
     val unit: String
 )
